requirements.txt
allure-pytest
allure-python-commons
playwright
pytest
python-slugify
requests